sm-user_interface_config/requirements.txt
cherrypy
paho-mqtt